@@ -4,10 +4,7 @@ import com.udacity.pricing.domain.price.Price;
 import com.udacity.pricing.service.PriceException;
 import com.udacity.pricing.service.PricingService;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 /**
@@ -32,4 +29,16 @@ public class PricingController {
         }
 
     }
+    //To remove the price assigned to the vehicle to be removed
+    @DeleteMapping ("/id")
+    public void delete(@RequestParam Long vehicleId) {
+        try {
+            PricingService.deletePrice(vehicleId);
+        } catch (PriceException ex) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Price Not Found", ex);
+        }
+
+    }
+
 }
