@@ -1,8 +1,5 @@
 package com.udacity.vehicles.api;
 
-import java.util.List;
-import java.util.stream.Collectors;
-import org.springframework.hateoas.VndErrors;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,6 +7,11 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+//import org.springframework.hateoas.VndErrors;
 
 /**
  * Implements the Error controller related to any errors handled by the Vehicles API
