@@ -2,20 +2,14 @@ package com.udacity.vehicles.domain.car;
 
 import com.udacity.vehicles.domain.Condition;
 import com.udacity.vehicles.domain.Location;
-import java.time.LocalDateTime;
-import javax.persistence.Embedded;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Transient;
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.*;
+import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
 
 /**
  * Declares the Car class, related variables and methods.
@@ -46,7 +40,7 @@ public class Car {
     @Embedded
     private Location location = new Location(0d, 0d);
 
-    @Transient
+//    @Transient
     private String price;
 
     public Long getId() {
