@@ -1,9 +1,13 @@
 package com.udacity.vehicles.service;
 
+import com.udacity.vehicles.client.maps.MapsClient;
+import com.udacity.vehicles.client.prices.PriceClient;
 import com.udacity.vehicles.domain.car.Car;
 import com.udacity.vehicles.domain.car.CarRepository;
-import java.util.List;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
 
 /**
  * Implements the car service create, read, update or delete
@@ -14,13 +18,19 @@ import org.springframework.stereotype.Service;
 public class CarService {
 
     private final CarRepository repository;
+    private final MapsClient mapsClient;
+    private final PriceClient priceClient;
 
-    public CarService(CarRepository repository) {
+    public CarService(CarRepository repository, MapsClient mapsClient, PriceClient priceClient) {
         /**
          * TODO: Add the Maps and Pricing Web Clients you create
          *   in `VehiclesApiApplication` as arguments and set them here.
          */
         this.repository = repository;
+        this.mapsClient = mapsClient;
+        this.priceClient = priceClient;
+
+
     }
 
     /**
@@ -42,7 +52,7 @@ public class CarService {
          *   If it does not exist, throw a CarNotFoundException
          *   Remove the below code as part of your implementation.
          */
-        Car car = new Car();
+
 
         /**
          * TODO: Use the Pricing Web client you create in `VehiclesApiApplication`
@@ -52,7 +62,9 @@ public class CarService {
          *   the pricing service each time to get the price.
          */
 
+        //TODO: Error happens here
 
+//        car.setPrice(priceClient.getPrice(id));
         /**
          * TODO: Use the Maps Web client you create in `VehiclesApiApplication`
          *   to get the address for the vehicle. You should access the location
@@ -61,9 +73,10 @@ public class CarService {
          * Note: The Location class file also uses @transient for the address,
          * meaning the Maps service needs to be called each time for the address.
          */
+//        Location location = car.getLocation();
+//        car.setLocation(mapsClient.getAddress(location));
+        return repository.findById(id).orElseThrow(CarNotFoundException::new);
 
-
-        return car;
     }
 
     /**
@@ -76,12 +89,26 @@ public class CarService {
             return repository.findById(car.getId())
                     .map(carToBeUpdated -> {
                         carToBeUpdated.setDetails(car.getDetails());
-                        carToBeUpdated.setLocation(car.getLocation());
+
+                        //Check if the new location changed or not
+                        //If not keep original location, otherwise retrieve another random location according to given lat and lon
+                        if(!car.getLocation().getLat().equals(carToBeUpdated.getLocation().getLat())
+                                || !car.getLocation().getLon().equals(carToBeUpdated.getLocation().getLon()))
+                            carToBeUpdated.setLocation(mapsClient.getAddress(car.getLocation()));
                         return repository.save(carToBeUpdated);
                     }).orElseThrow(CarNotFoundException::new);
-        }
+        }else {
+            return Optional.of(car).map(carToBeSaved->{
+                carToBeSaved.setDetails(car.getDetails());
 
-        return repository.save(car);
+                carToBeSaved.setLocation(mapsClient.getAddress(car.getLocation()));
+                //Add price for the car to be created
+                Car carWithId = repository.save(carToBeSaved);
+                carWithId.setPrice(priceClient.getPrice(carWithId.getId()));
+                return repository.save(carWithId);
+            }).orElseThrow(CarNotFoundException::new);
+
+        }
     }
 
     /**
@@ -93,12 +120,13 @@ public class CarService {
          * TODO: Find the car by ID from the `repository` if it exists.
          *   If it does not exist, throw a CarNotFoundException
          */
-
+        Car car = repository.findById(id).orElseThrow(CarNotFoundException::new);
 
         /**
          * TODO: Delete the car from the repository.
          */
-
+        repository.delete(car);
+        priceClient.deletePrice(car.getId());
 
     }
 }
